+++ lib/views/app_locks/security_questions/security_questions_view.dart
@@ -14,9 +14,6 @@ class SecurityQuestionsRoute extends BaseRoute {
   SecurityQuestionsRoute();
 
   @override
-  bool get preferredNestedRoute => true;
-
-  @override
   Widget buildPage(BuildContext context) => SecurityQuestionsView(params: this);
 }
