+++ lib/views/home/home_view_model.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:storypad/core/objects/search_filter_object.dart';
 import 'package:storypad/core/storages/search_filter_storage.dart';
 import 'package:storypad/views/search/filter/search_filter_view.dart';
-import 'package:storypad/widgets/sheets/search_filter_sheet.dart';
+import 'package:storypad/widgets/bottom_sheets/search_filter_bottom_sheet.dart';
 import 'package:storypad/widgets/view/base_view_model.dart';
 import 'package:storypad/core/databases/models/collection_db_model.dart';
 import 'package:storypad/core/databases/models/preference_db_model.dart';
@@ -16,7 +16,7 @@ import 'package:storypad/core/services/backups/restore_backup_service.dart';
 import 'package:storypad/core/storages/new_stories_count_storage.dart';
 import 'package:storypad/core/types/path_type.dart';
 import 'package:storypad/providers/backup_provider.dart';
-import 'package:storypad/widgets/sheets/nickname_bottom_sheet.dart';
+import 'package:storypad/widgets/bottom_sheets/nickname_bottom_sheet.dart';
 import 'package:storypad/views/stories/edit/edit_story_view.dart';
 import 'package:storypad/views/stories/show/show_story_view.dart';
 
@@ -111,7 +111,7 @@ class HomeViewModel extends BaseViewModel {
   }
 
   Future<void> goToFilter(BuildContext context, {bool save = false}) async {
-    final result = await SearchFilterSheet(
+    final result = await SearchFilterBottomSheet(
       params: SearchFilterRoute(
         initialTune: currentSearchFilter,
         resetTune: initialSearchFilter,
@@ -143,11 +143,7 @@ class HomeViewModel extends BaseViewModel {
   }
 
   void changeName(BuildContext context) async {
-    final result = await NicknameBottomSheet(nickname: nickname).show(
-      context: context,
-      isScrollControlled: true,
-    );
-
+    final result = await NicknameBottomSheet(nickname: nickname).show(context: context);
     if (result is String) {
       PreferenceDbModel.db.nickname.set(result);
       nickname = PreferenceDbModel.db.nickname.get();
