+++ lib/views/onboarding/steps/onboarding_hello_view.dart
@@ -9,9 +9,6 @@ class OnboardingHelloRoute extends BaseRoute {
     required this.nickname,
   });
 
-  @override
-  bool get preferredNestedRoute => true;
-
   final String nickname;
 
   @override
