+++ lib/views/onboarding/steps/step_1/onboarding_step_1_view.dart
@@ -18,9 +18,6 @@ class OnboardingStep1Route extends BaseRoute {
   OnboardingStep1Route();
 
   @override
-  bool get preferredNestedRoute => true;
-
-  @override
   Widget buildPage(BuildContext context) => OnboardingStep1View(params: this);
 }
