+++ lib/views/onboarding/steps/step_3/onboarding_step_3_view.dart
@@ -20,9 +20,6 @@ class OnboardingStep3Route extends BaseRoute {
   OnboardingStep3Route();
 
   @override
-  bool get preferredNestedRoute => true;
-
-  @override
   Widget buildPage(BuildContext context) => OnboardingStep3View(params: this);
 }
