+++ lib/views/onboarding/steps/step_4/onboarding_step_4_view.dart
@@ -12,9 +12,6 @@ class OnboardingStep4Route extends BaseRoute {
   OnboardingStep4Route();
 
   @override
-  bool get preferredNestedRoute => true;
-
-  @override
   Widget buildPage(BuildContext context) => OnboardingStep4View(params: this);
 }
