+++ lib/widgets/bottom_sheets/nickname_bottom_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:storypad/widgets/sheets/base_bottom_sheet.dart';
+import 'package:storypad/widgets/bottom_sheets/base_bottom_sheet.dart';
 import 'package:storypad/widgets/sp_default_text_controller.dart';
 
 class NicknameBottomSheet extends BaseBottomSheet {
